Flex hero banner, align it with header
Flexed the hero banner in a row format. added padding to the left
section and the header logo so they would line up.
flexed the two sub div's in hero banner.
encased submit button html in a div so it would not stretch.
added sizing to image so it would stay one size.